+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Handling of standard and JUnitParams parameters without `ArtifactParams` interface
 ### Fixed
 - 'retrieveLeaf' and 'stopRunner' methods do not start new launch in case of empty suites / tests
 - 'stopLaunch' method does not start new launch in case of there was no any tests passed